Set turn to `null` rather than relying on a dummy combatant
Foundry v9 allows the turn to be set to null in updates. This allows for
no combatant to be active on round change, obviating the need for a
dummy combatant to absorb the 0th turn.

With no need for a dummy combatant, there's no need for the function to
add a missing dummy combatant. The function remains, but is now NOP and
marked as deprecated.